backend/requirements.txt
# Flask and web framework dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Machine Learning and AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
huggingface-hub>=0.18.0

# Vector Database
qdrant-client>=1.6.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# HTTP and networking
requests>=2.31.0
httpx>=0.25.0

# AWS Services
boto3>=1.28.0
botocore>=1.31.0

# Development and utilities
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For production deployment
gunicorn>=21.2.0
gevent>=23.9.0